Split Python wrapper from VQE specific code
